backend/requirements.txt
openai>=1.12.0
qdrant-client>=1.7.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.0.0
python-multipart>=0.0.6
